+++ src/main/java/com/github/mc1arke/sonarqube/plugin/almclient/azuredevops/model/ConnectionData.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (C) 2024 Michael Clarke
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ */
+package com.github.mc1arke.sonarqube.plugin.almclient.azuredevops.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class ConnectionData {
+
+    private final Identity authenticatedUser;
+
+    public ConnectionData(@JsonProperty("authenticatedUser") Identity authenticatedUser) {
+        this.authenticatedUser = authenticatedUser;
+    }
+
+    public Identity getAuthenticatedUser() {
+        return authenticatedUser;
+    }
+
+    public static class Identity {
+
+        private final String id;
+
+        public Identity(@JsonProperty("id") String id) {
+            this.id = id;
+        }
+
+        public String getId() {
+            return id;
+        }
+    }
+
+}
